requirements.txt
Django
boto3
python-dotenv
